combined all qwen_omni modules into one file
